Refactor code and add abstraction
Quite frankly, this commit is too big to sort.
It takes the existing controve files and changes
the package structure to make it simpler. It also
adds an abstraction layer as well as building the
base framework for Controve.java as an active component
as well as Conductor spawner for running threads.

This commit is 100% breaking.